style: move sidebar icons to Icons
