+++ wondrous-app/components/Common/Sidebar/Entity/AboutEntity/EntityMenu/styles.tsx
@@ -1,5 +1,5 @@
 import { ButtonBase, Menu, MenuItem, Typography } from '@mui/material';
-import PodsIcon from 'components/Common/Sidebar/Common/icons/pods.svg';
+import PodsIcon from 'components/Icons/Sidebar/pods.svg';
 import Arrow from 'components/Icons/arrow.svg';
 import { DAOIcon } from 'components/Icons/dao';
 import styled from 'styled-components';
