+++ wondrous-app/components/Common/Sidebar/Entity/List.tsx
@@ -1,13 +1,13 @@
 import { useQuery } from '@apollo/client';
-import CheckBoxIcon from 'components/Common/Sidebar/Common/icons/checkBox.svg';
-import ContentPaste from 'components/Common/Sidebar/Common/icons/contentPaste.svg';
-import FlagIcon from 'components/Common/Sidebar/Common/icons/flag.svg';
-import FolderIcon from 'components/Common/Sidebar/Common/icons/folder.svg';
-import GroupIcon from 'components/Common/Sidebar/Common/icons/group.svg';
-import PieChartIcon from 'components/Common/Sidebar/Common/icons/pieChart.svg';
-import ShowChartIcon from 'components/Common/Sidebar/Common/icons/showChart.svg';
-import StackIcon from 'components/Common/Sidebar/Common/icons/stack.svg';
-import StartIcon from 'components/Common/Sidebar/Common/icons/star.svg';
+import CheckBoxIcon from 'components/Icons/Sidebar/checkBox.svg';
+import ContentPaste from 'components/Icons/Sidebar/contentPaste.svg';
+import FlagIcon from 'components/Icons/Sidebar/flag.svg';
+import FolderIcon from 'components/Icons/Sidebar/folder.svg';
+import GroupIcon from 'components/Icons/Sidebar/group.svg';
+import PieChartIcon from 'components/Icons/Sidebar/pieChart.svg';
+import ShowChartIcon from 'components/Icons/Sidebar/showChart.svg';
+import StackIcon from 'components/Icons/Sidebar/stack.svg';
+import StartIcon from 'components/Icons/Sidebar/star.svg';
 import { Label, ListWrapper } from 'components/Common/Sidebar/Common/styles';
 import Item from 'components/Common/SidebarItem';
 import { GET_ORG_SIDEBAR_COUNT, GET_POD_SIDEBAR_COUNT } from 'graphql/queries';
